fw-5326, move helper functions into access factory
