+++ firstvoices/backend/tests/factories/access.py
@@ -4,8 +4,9 @@ from django.contrib.auth.models import AnonymousUser
 from factory.django import DjangoModelFactory
 
 from backend.models.app import AppMembership
-from backend.models.constants import AppRole
+from backend.models.constants import AppRole, Role, Visibility
 from backend.models.sites import Language, LanguageFamily, Membership, Site
+from backend.tests import factories
 
 
 class AnonymousUserFactory(DjangoModelFactory):
@@ -92,3 +93,37 @@ def get_app_admin(role):
 
 def get_superadmin():
     return get_app_admin(AppRole.SUPERADMIN)
+
+
+def get_member_of_other_site():
+    _, user = get_site_with_member(Visibility.PUBLIC, Role.LANGUAGE_ADMIN)
+    return user
+
+
+def get_site_with_authenticated_member(
+    client, visibility=Visibility.PUBLIC, role=Role.MEMBER
+):
+    site, user = factories.get_site_with_member(visibility, role)
+    client.force_authenticate(user=user)
+    return site, user
+
+
+def get_site_with_authenticated_nonmember(client, visibility=Visibility.PUBLIC):
+    site = factories.SiteFactory.create(visibility=visibility)
+    user = factories.get_non_member_user()
+    client.force_authenticate(user=user)
+    return site, user
+
+
+def get_site_with_staff_user(client=None, visibility=Visibility.PUBLIC):
+    site = factories.SiteFactory.create(visibility=visibility)
+    user = factories.get_app_admin(AppRole.STAFF)
+    client.force_authenticate(user=user)
+    return site, user
+
+
+def get_site_with_anonymous_user(client=None, visibility=Visibility.PUBLIC):
+    # client is intentionally ignored so all these site_with_user functions can have the same signature
+    site = factories.SiteFactory.create(visibility=visibility)
+    user = factories.get_anonymous_user()
+    return site, user
